Refactor hash and token apis under internal/crypto, leave aliases
